This will need a lot of updates right now
This will probably be deleted, just meant to save my progress in this for now. This might as well be garbage too. This simulator aims to use buffers and totally removes the dependency of each stage from global registers. Needed for working in pipeline. When we run reverse loop W M E D F, we will have all the required buffers being updated and appropriate buffers being used.